backend/requirements.txt
azure-core==1.31.0
azure-storage-blob==12.23.0
flasgger==0.9.7.1
Flask==3.0.3
Flask_Cors==5.0.0
Flask_JWT_Extended==4.6.0
flask_marshmallow==1.2.1
Flask_PyMongo==2.3.0
marshmallow==3.22.0
openai==1.47.0
PyJWT==2.9.0
pymongo==4.9.1
python-dotenv==1.0.1
bcrypt==4.2.0
sendgrid==6.11.0
twilio==9.3.1
gunicorn~=20.1.0
flask[async]
flask-restx
pytest
google-auth